+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- Support new PO Token context requested by yt-dlp.
+- Support new PO Token context requested by yt-dlp. By default, providers only support "gvs" PO Token context for backwards compatibility.
 
 ## [0.2.0]